rename components with naming guide
