+++ src/components/UserInviteModal.tsx
@@ -3,25 +3,23 @@ import { useRouter as useNextRouter } from 'next/router';
 import tinykeys from 'tinykeys';
 
 import { routes, useRouter } from '../hooks/router';
-import { createProjectKeys, createHotkeys } from '../utils/hotkeys';
+import { createHotkeys, inviteUserKeys } from '../utils/hotkeys';
 import { DialogModal } from './DialogModal';
-import { CreateProject } from './CreateProject';
+import { UserInviteForm } from './UserInviteForm';
 
-export const CreateProjectModal = () => {
+export const UserInviteModal = () => {
     const nextRouter = useNextRouter();
     const router = useRouter();
     const [modalVisible, setModalVisibility] = useState(false);
-    const isCreateProjectPath = nextRouter.pathname === routes.createProject();
-    const showModalOrNavigate = (navigate: () => void) => (isCreateProjectPath ? navigate() : setModalVisibility(true));
+    const isInviteUsersPath = nextRouter.pathname === routes.inviteUsers();
+    const showModalOrNavigate = (navigate: () => void) => (isInviteUsersPath ? navigate() : setModalVisibility(true));
     const onModalClose = useCallback(() => setModalVisibility(false), [setModalVisibility]);
 
-    useEffect(() =>
-        tinykeys(window, createHotkeys([createProjectKeys, () => showModalOrNavigate(router.createProject)])),
-    );
+    useEffect(() => tinykeys(window, createHotkeys([inviteUserKeys, () => showModalOrNavigate(router.inviteUsers)])));
 
     return (
         <DialogModal visible={modalVisible} onClose={onModalClose}>
-            <CreateProject onCreate={(slug) => slug && router.project(slug)} />
+            <UserInviteForm onCreate={() => setModalVisibility(false)} />
         </DialogModal>
     );
 };
